backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
deprecated==1.2.18
dnspython==2.7.0
dnstwist==20240812
flask==3.1.0
flask-cors==5.0.0
frozenlist==1.5.0
geoip2==4.8.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-pubsub==2.28.0
googleapis-common-protos==1.67.0
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
maxminddb==2.6.3
multidict==6.1.0
mysql-connector-python==9.2.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
pip==24.3.1
ppdeep==20200505
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
py-tlsh==4.7.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlalchemy==2.0.38
sqlmodel==0.0.22
tld==0.13
typing-extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
